move Multichain API tests out of flask/ directory
