+++ test/e2e/multichain-api/wallet_getSession.spec.ts
@@ -1,8 +1,8 @@
 import { strict as assert } from 'assert';
-import { unlockWallet, withFixtures } from '../../helpers';
-import FixtureBuilder from '../../fixture-builder';
-import { DEFAULT_FIXTURE_ACCOUNT } from '../../constants';
-import TestDappMultichain from '../../page-objects/pages/test-dapp-multichain';
+import { unlockWallet, withFixtures } from '../helpers';
+import FixtureBuilder from '../fixture-builder';
+import { DEFAULT_FIXTURE_ACCOUNT } from '../constants';
+import TestDappMultichain from '../page-objects/pages/test-dapp-multichain';
 import {
   DEFAULT_MULTICHAIN_TEST_DAPP_FIXTURE_OPTIONS,
   getExpectedSessionScope,
