+++ test/e2e/multichain-api/wallet_notify.spec.ts
@@ -1,7 +1,7 @@
 import { strict as assert } from 'assert';
-import { unlockWallet, withFixtures } from '../../helpers';
-import FixtureBuilder from '../../fixture-builder';
-import TestDappMultichain from '../../page-objects/pages/test-dapp-multichain';
+import { unlockWallet, withFixtures } from '../helpers';
+import FixtureBuilder from '../fixture-builder';
+import TestDappMultichain from '../page-objects/pages/test-dapp-multichain';
 import {
   DEFAULT_MULTICHAIN_TEST_DAPP_FIXTURE_OPTIONS,
   type FixtureCallbackArgs,
