+++ test/e2e/multichain-api/wallet_revokeSession.spec.ts
@@ -7,9 +7,9 @@ import {
   unlockWallet,
   WINDOW_TITLES,
   withFixtures,
-} from '../../helpers';
-import FixtureBuilder from '../../fixture-builder';
-import TestDappMultichain from '../../page-objects/pages/test-dapp-multichain';
+} from '../helpers';
+import FixtureBuilder from '../fixture-builder';
+import TestDappMultichain from '../page-objects/pages/test-dapp-multichain';
 import {
   DEFAULT_MULTICHAIN_TEST_DAPP_FIXTURE_OPTIONS,
   addAccountInWalletAndAuthorize,
