+++ test/e2e/multichain-api/wallet_sessionChanged.spec.ts
@@ -6,10 +6,10 @@ import {
   unlockWallet,
   WINDOW_TITLES,
   withFixtures,
-} from '../../helpers';
-import { Driver } from '../../webdriver/driver';
-import FixtureBuilder from '../../fixture-builder';
-import TestDappMultichain from '../../page-objects/pages/test-dapp-multichain';
+} from '../helpers';
+import { Driver } from '../webdriver/driver';
+import FixtureBuilder from '../fixture-builder';
+import TestDappMultichain from '../page-objects/pages/test-dapp-multichain';
 import {
   addAccountInWalletAndAuthorize,
   DEFAULT_MULTICHAIN_TEST_DAPP_FIXTURE_OPTIONS,
