+++ test/e2e/solana-wallet-standard/connect.spec.ts
@@ -6,7 +6,7 @@ import {
 } from '../solana/common-solana';
 import { TestDappSolana } from '../../page-objects/pages/test-dapp-solana';
 import { regularDelayMs, WINDOW_TITLES } from '../../helpers';
-import { updateNetworkCheckboxes } from '../multichain-api/testHelpers';
+import { updateNetworkCheckboxes } from '../../multichain-api/testHelpers';
 import {
   account1Short,
   account2Short,
